# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/refunds/eligibility": {
      "get": {
        "operationId": "Refunds_eligibility",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** CUSTOMER"
      }
    },
    "/api/v1/refunds": {
      "post": {
        "operationId": "Refunds_request",
        "summary": "",
        "description": "Ask for a refund. The organizer decides; you're emailed either way.\n\n**Roles:** CUSTOMER",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/refunds/mine": {
      "get": {
        "operationId": "Refunds_mine",
        "summary": "",
        "description": "Your refunds and requests.\n\n**Roles:** CUSTOMER",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/refunds/{id}/withdraw": {
      "post": {
        "operationId": "Refunds_withdraw",
        "summary": "",
        "description": "Take back a request that hasn't been decided yet.\n\n**Roles:** CUSTOMER",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/events/{id}/refunds": {
      "get": {
        "operationId": "Refunds_listForEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "REQUESTED",
                "APPROVED",
                "REJECTED",
                "PROCESSED",
                "WITHDRAWN"
              ],
              "type": "string"
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PROVIDER",
                "MANUAL"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      },
      "post": {
        "operationId": "Refunds_refundTickets",
        "summary": "",
        "description": "Refund tickets directly (no request needed). Tickets stop working immediately.\n\n**Roles:** ORGANIZER, ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundTicketsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/refunds/{id}/approve": {
      "post": {
        "operationId": "Refunds_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      }
    },
    "/api/v1/refunds/{id}/reject": {
      "post": {
        "operationId": "Refunds_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      }
    },
    "/api/v1/admin/refunds": {
      "get": {
        "operationId": "Refunds_adminList",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "REQUESTED",
                "APPROVED",
                "REJECTED",
                "PROCESSED",
                "WITHDRAWN"
              ],
              "type": "string"
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PROVIDER",
                "MANUAL"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      }
    },
    "/api/v1/admin/refunds/{id}/mark-paid": {
      "post": {
        "operationId": "Refunds_markPaid",
        "summary": "",
        "description": "A manual refund has been paid back: record the reference and tell the customer.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkPaidDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/refunds/{id}/retry": {
      "post": {
        "operationId": "Refunds_retry",
        "summary": "",
        "description": "Try a failed provider refund again.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/refunds/run": {
      "post": {
        "operationId": "Refunds_run",
        "summary": "",
        "description": "Send due provider refunds now.\n\n**Roles:** ADMIN",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/payments/{id}/refund": {
      "post": {
        "operationId": "Refunds_refundPayment",
        "summary": "",
        "description": "Refund one payment in full: all its tickets and the booking fee (Phase 6 endpoint, now real).\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/events/{id}/refund-all": {
      "post": {
        "operationId": "Refunds_refundAll",
        "summary": "",
        "description": "Refund everyone for a cancelled event whose organizer chose to handle refunds.\n\n**Roles:** ADMIN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Refunds"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Event Ticketing Platform API",
    "description": "REST API shared by the web app and the organizer/staff mobile apps. Amounts are integer minor units (butut, GMD). Sign in with POST /api/v1/auth/login, then send the access token as `Authorization: Bearer <token>`. See docs/api.md.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestRefundDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "ticketIds": {
            "description": "Tickets to refund; omit for all of your valid tickets in the order",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "orderId"
        ]
      },
      "RefundTicketsDto": {
        "type": "object",
        "properties": {
          "ticketIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          },
          "includeFee": {
            "type": "boolean",
            "description": "Admin only: also return the booking fee"
          }
        },
        "required": [
          "ticketIds"
        ]
      },
      "DecisionDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "Shown to the customer. Required when rejecting.",
            "maxLength": 500
          }
        }
      },
      "MarkPaidDto": {
        "type": "object",
        "properties": {
          "reference": {
            "type": "string",
            "description": "Reference of the money sent back (bank transfer / Wave payout)",
            "maxLength": 200
          }
        },
        "required": [
          "reference"
        ]
      },
...
    }
  }
}